Fix zoom when on page > 1 bug & disable server select on refresh.
This commit fixes a couple of issues. If you zoom in on an account
when on a page other than the first, it keeps thinking you want to
swap to this page when you are in the zoom view, causing out of bounds
exception. Fixed.

The dropdown menu to select servers was active during refresh. If you
manage to switch server context during refresh, this will cause the
application to crash. Fixed.