Make it possible to cancel ongoing import follower / following
There is a button to cancel import, this now works.

This is probably not done in a proper way, but as this will be dealt
with when I try to figure out how jetpack compose is supposed to handle
separate threads / coroutines with active UI updates.

For now, the most important part is, it works :P