Resizeable UI (#16)
A lot of tidy-up work done in this commit. Views have been decoupled, state objects have been 
drastically reduced, and state moved as closed to usage as possible. UI is made to be more 
event driven than it was. Most all the work in this commit adds no new functionalities, it is only
a needed refactoring.

Initial support for UI resizing is added, when you resize the UI, the amount of rows showing 
follower / followings is changed. This is possibly the most needed resize functionality, as it enables
you to view more followers pr page.

Initial testing has been done, but more testing is needed before release next weekend, so use with 
care.